Lift the state into the Board.
A common pattern for handling communication between components is to
lift the state of the components into a higher level container
component. In this case we lift the state of each Square into the
containing Board component. The reason for this is that it is generally
easier to control sub components from a higher level component rather
than keeping state in the sub components themselves.