Lift board state up to the Game component.
We want the Game to show a history of moves and allow the player to
click on these moves to transition the board state to what it was at
that point of the move. In order to do so the Game needs to track the
history and instruct the board to go to a specific state. Thus, we lift
the state of the squares up into the Game.